Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pytesseract
pillow
pymupdf
python-multipart
fhir.resources
pydicom
celery
redis
